fastapi-skeleton-template/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiofiles==24.1.0
aiohttp==3.10.5
alembic==1.13.2
Authlib==1.3.1
colorama==0.4.6
fastapi==0.111.1
icecream==2.1.3
loguru==0.7.2
multidict==6.0.5
scipy==1.11.3  # numpy 1.26.4
openpyxl==3.1.5
orjson==3.10.7
pandas==2.2.2
pathlib2==2.3.7.post1
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pydantic==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.8
sendgrid==6.11.0
SQLAlchemy==2.0.31
starlette==0.37.2
uvicorn==0.30.3
aioredis==2.0.1  # async-timeout==4.0.3
funboost==46.2
async_timeout==4.0.3
pydantic-settings==2.5.2
werkzeug==3.0.4
asyncpg==0.29.0
